# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HillSpace API",
    "description": "Estate management backend — auth, listings, search, verification, escrow",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" }
        }
      },
      "LocationDto": {
        "type": "object",
        "required": ["address", "city", "state"],
        "properties": {
          "address": { "type": "string", "example": "12 Admiralty Way" },
          "city": { "type": "string", "example": "Lagos" },
          "state": { "type": "string", "example": "Lagos" },
          "country": { "type": "string", "example": "Nigeria" },
          "lat": { "type": "number" },
          "lng": { "type": "number" }
        }
      },
      "CreateListingDto": {
        "type": "object",
        "required": ["title", "description", "propertyType", "purpose", "price", "location"],
        "properties": {
          "title": { "type": "string", "example": "4-bed duplex in Lekki" },
          "description": { "type": "string" },
          "propertyType": {
            "type": "string",
            "enum": ["apartment", "house", "land", "commercial", "duplex", "bungalow"]
          },
          "purpose": { "type": "string", "enum": ["sale", "rent"] },
          "price": { "type": "number", "example": 85000000 },
          "currency": { "type": "string", "example": "NGN" },
          "bedrooms": { "type": "number", "example": 4 },
          "bathrooms": { "type": "number", "example": 5 },
          "areaSqm": { "type": "number", "example": 320 },
          "location": { "$ref": "#/components/schemas/LocationDto" },
          "amenities": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["parking", "security"]
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "under_offer", "sold", "rented", "archived"]
          }
        }
      },
      "UpdateListingDto": {
        "allOf": [{ "$ref": "#/components/schemas/CreateListingDto" }]
      },
...
      "ReviewVerificationDto": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "enum": ["approved", "rejected"] },
          "reviewNotes": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "security": [{ "access-token": [] }],
        "responses": {
          "200": { "description": "Current user" }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update current user profile",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated user" }
        }
      }
    },
    "/api/listings": {
      "get": {
        "tags": ["Listings"],
        "summary": "Search and filter listings",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          {
            "name": "propertyType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["apartment", "house", "land", "commercial", "duplex", "bungalow"]
            }
          },
          {
            "name": "purpose",
            "in": "query",
            "schema": { "type": "string", "enum": ["sale", "rent"] }
          },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "state", "in": "query", "schema": { "type": "string" } },
          { "name": "minPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "bedrooms", "in": "query", "schema": { "type": "number" } },
          { "name": "bathrooms", "in": "query", "schema": { "type": "number" } },
          { "name": "page", "in": "query", "schema": { "type": "number", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "number", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated listings" }
        }
      },
      "post": {
        "tags": ["Listings"],
        "summary": "Create a listing",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateListingDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Listing created" }
        }
      }
    },
    "/api/listings/{id}": {
      "get": {
        "tags": ["Listings"],
        "summary": "Get listing by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Listing" }
        }
      },
      "patch": {
        "tags": ["Listings"],
        "summary": "Update a listing",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateListingDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated listing" }
        }
      },
      "delete": {
        "tags": ["Listings"],
        "summary": "Delete a listing",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted" }
        }
      }
    },
    "/api/listings/{id}/images": {
      "post": {
        "tags": ["Listings"],
        "summary": "Upload listing images",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Images uploaded" }
        }
      }
    },
    "/api/listings/{id}/publish": {
      "post": {
        "tags": ["Listings"],
        "summary": "Publish a listing (set status active)",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "201": { "description": "Published" }
        }
      }
    },
    "/api/verification": {
      "post": {
        "tags": ["Verification"],
        "summary": "Submit KYC, agent, or listing verification docs",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["type", "documents"],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["kyc", "agent", "listing"]
                  },
                  "listingId": { "type": "string" },
                  "notes": { "type": "string" },
                  "documents": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Verification submitted" }
        }
      }
    },
    "/api/verification/me": {
      "get": {
        "tags": ["Verification"],
        "summary": "List my verification requests",
        "security": [{ "access-token": [] }],
        "responses": {
          "200": { "description": "My requests" }
        }
      }
    },
    "/api/verification/pending": {
      "get": {
        "tags": ["Verification"],
        "summary": "List pending verification requests (admin)",
        "security": [{ "access-token": [] }],
        "responses": {
          "200": { "description": "Pending requests" }
        }
      }
    },
    "/api/verification/{id}/review": {
      "patch": {
        "tags": ["Verification"],
        "summary": "Approve or reject a verification request (admin)",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewVerificationDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Reviewed" }
        }
      }
    },
...
  }
}